Refactor Server to Leverage Docker Compose
In anticipation of additional containers, this commit refactors the MCP Server into its own directory to be run as one of several containers.

* Create Docker Compose file for configuring and launching multiple containers
* Update the README to include new setup instructions
* Update the README with examples of leveraging new Docker environment to execute tool calls
* Move all MCP Server code to a new directory, mcp-server
* Update GitHub Actions to use the mcp-server directory for testing
* Update GitHub Build Action to create and verify Docker images
* Create mcp-connect script to check if Docker containers have been started already before attempting to start (useful for LLM Configuration)
* Remove npm caching option in Actions - this causes failures due to no root package.lock.json
* Pass ENV Variable defined in MCP Server Configuration to Docker via the mcp-connect bash script
* Update README to include LLM Configuration Example
