+++ README.md
@@ -8,35 +8,55 @@ A [Model Context Protocol (MCP)](https://modelcontextprotocol.io/introduction) s
 
 ## Requirements
 * A valid Census Bureau [API key](https://api.census.gov/data/key_signup.html)
-* Docker (e.g. Docker Desktop)
-* Node 16+
+* Docker (i.e. Docker Desktop)
+* Node 18+
 
-## Setup
+## Using the MCP Server
 
-Run `npm install` to install the project’s dependencies.
+After saving the project locally, configure your LLM Client to leverage the Census API MCP Server via your LLM Client’s configuration file. The syntax for the configuration file may vary by LLM. Here is an example configuration file that includes the appropriate scripts for launching the server:
 
-Then run `npm run build` to build the project.
+```
+{
+  "mcpServers": {
+    "mcp-census-api": {
+      "command": "bash",
+      "args": [
+        "/Path/To/Server/mcp-server-census-api/scripts/mcp-connect.sh"
+      ],
+      "env": {
+        "CENSUS_API_KEY": "YOUR_CENSUS_API_KEY"
+      }
+    }
+  }
+}
+```
 
-With Docker desktop running, run `npm run docker-build` to build the project in Docker.
+Note that the `CENSUS_API_KEY` variable is required. This defines the `env` variable in the Client and passes it to the MCP server via the `mcp-connect` script.
+
+Be sure to update the path to the `mcp-server-census-api` in `args` and update the `CENSUS_API_KEY` to a valid API key.
 
 ## Development
 
-By default, all logging functions are disabled to prevent `json` validation errors when interacting with the MCP server through MCP clients. To enable logging for development purposes, set `DEBUG_LOGS=true` when running the server in Docker, e.g. `-e DEBUG_LOGS=true`. 
+Run `docker compose` from the root of the project to build the containers.
+
+By default, all logging functions are disabled in the `mcp-server` to prevent `json` validation errors when interacting with the MCP server through MCP clients. To enable logging for development purposes, set `DEBUG_LOGS=true` when interacting with the server directly using the examples below, e.g. `echo '{CALL_ARGUMENTS}' docker exec -e DEBUG_LOGS=true -i mcp-server node dist/index.js`. 
 
 ### Testing
 
-This project uses [Vitest](https://vitest.dev/) to test the functionality of the server.
+This project uses [Vitest](https://vitest.dev/) to test the server.
 
-Prior to running tests, a valid Census Bureau [API key](https://api.census.gov/data/key_signup.html) is required. This key should be defined in the `.env` file of the root directory of the project. The `sample.env` offers an example of how this `.env` file should look.
+Prior to running tests, a valid Census Bureau [API key](https://api.census.gov/data/key_signup.html) is required. This key should be defined in the `.env` file of the `mcp-server` directory. The `sample.env` offers an example of how this `.env` file should look.
 
 To run tests, run `npm run test`. To run ESLint, run `npm run lint`.
 
-## Key Concepts
+## MCP Server Architecture
 
-* `index.ts` - Starts the MCP Server and registers tools.
-* `server.ts` - Defines the `McpServer` class that handles calls to the server, e.g. how `tools/list` and `tools/calls` respond to requests
-* `tools/` - Includes tool definitions and shared classes, e.g. `BaseTool` and `ToolRegistry`, to reduce repetition and exposes the tools list to the server
-* `schema/` - Houses each tool’s schema and is used to validate schemas in tests
+* `docker-compose` - Organizes the project into different containers.
+* `mcp-server` - Source code for the MCP Server.
+* `mcp-server/index.ts` - Starts the MCP Server and registers tools.
+* `mcp-server/server.ts` - Defines the `McpServer` class that handles calls to the server, e.g. how `tools/list` and `tools/calls` respond to requests
+* `mcp-server/tools/` - Includes tool definitions and shared classes, e.g. `BaseTool` and `ToolRegistry`, to reduce repetition and exposes the tools list to the server
+* `mcp-server/schema/` - Houses each tool’s schema and is used to validate schemas in tests
 
 ## Available Methods
 
@@ -46,10 +66,9 @@ The MCP server exposes two methods: `tools/list` and `tools/call`.
 
 To list available tools, use the `tools/list` method with no arguments. `tools/list` is a standard method that is often called by LLMs when the client is initialized.
 
-### Example
+#### How to Run Outside an LLM
 ```
-echo '{"jsonrpc":"2.0","id":1,"method":"tools/list"}' \
-| docker run --rm -i census-api
+echo '{"jsonrpc":"2.0","id":1,"method":"tools/list"}' | docker exec -i mcp-server node dist/index.js
 ```
 
 ## Available Tools
@@ -60,10 +79,10 @@ The `describe-dataset` tool is used for fetching metadata about a given dataset 
 * Dataset (Required) - The identifier of the dataset, e.g. `'acs/acs1'`
 * Year (Optional) - The vintage of the dataset, e.g. `1987`
 
-#### Example
+#### How to Run Outside an LLM
 ```
 echo '{"jsonrpc":"2.0","id":1,"method":"tools/call","params":{"name":"describe-dataset","arguments":{"dataset":"acs/acs1","year":2022}}}' \
-| docker run --rm -i -e CENSUS_API_KEY=YOUR_API_KEY census-api
+| docker exec -i mcp-server node dist/index.js
 ```
 
 ### Fetch Dataset Geography
@@ -71,10 +90,10 @@ The `fetch-dataset-geography` tool is used for fetching available geography leve
 * Dataset (Required) - The identifier of the dataset, e.g. `'acs/acs1'`
 * Year (Optional) - The vintage of the dataset, e.g. `1987`
 
-#### Example
+#### How to Run Outside an LLM
 ```
 echo '{"jsonrpc":"2.0","id":1,"method":"tools/call","params":{"name":"fetch-dataset-geography","arguments":{"dataset":"acs/acs1","year":2022}}}' \
-| docker run --rm -i -e CENSUS_API_KEY=YOUR_API_KEY census-api
+| docker exec -i mcp-server node dist/index.js
 ```
 
 ### Fetch Dataset Variables
@@ -83,10 +102,10 @@ The `fetch-dataset-variables` tool is used for fetching variables for filtering 
 * Group (Optional) - Filter variables by a specific group for this dataset, e.g. `'S0101'`
 * Year (Optional) - The vintage of the dataset, e.g. `1987`
 
-#### Example
+#### How to Run Outside an LLM
 ```
 echo '{"jsonrpc":"2.0","id":1,"method":"tools/call","params":{"name":"fetch-dataset-variables","arguments":{"dataset":"acs/acs1","year":2022}}}' \
-| docker run --rm -i -e CENSUS_API_KEY=YOUR_API_KEY census-api
+| docker exec -i mcp-server node dist/index.js
 ```
 
 ### Fetch Summary Table
@@ -102,10 +121,10 @@ The `fetch-summary-table` tool is used for fetching a summary table from the Cen
 * Predicates (Optional) - Filter options for the dataset, e.g. `'for': 'state*'`
 * Descriptive (Optional) - Adds variable labels to API response (default: `false`), e.g. `true`
 
-#### Example
+#### How to Run Outside an LLM
 ```
 echo '{"jsonrpc":"2.0","id":1,"method":"tools/call","params":{"name":"fetch-summary-table","arguments":{"dataset":"acs/acs1","year":2022, "get": { "variables":["NAME","B01001_001E"] },"for":"state:01,13"}}}' \
-| docker run --rm -i -e CENSUS_API_KEY=YOUR_API_KEY census-api
+| docker exec -i mcp-server node dist/index.js
 ```
 
 ## Resources
