+++ mcp-server/eslint.config.js
