+++ mcp-server/sample.env
