+++ mcp-server/src/schema/dataset-variables.schema.ts
