+++ mcp-server/src/schema/describe-dataset.schema.ts
