+++ mcp-server/src/schema/summary-table.schema.ts
