+++ mcp-server/src/schema/table.schema.ts
