+++ mcp-server/src/tools/base.tool.ts
