+++ mcp-server/src/tools/describe-dataset.tool.ts
