+++ mcp-server/src/tools/fetch-dataset-geography.tool.ts
