+++ mcp-server/src/tools/fetch-dataset-variables.tool.ts
