+++ mcp-server/src/tools/fetch-summary-table.tool.ts
