+++ mcp-server/src/types/base.types.ts
