+++ mcp-server/tsconfig.json
